chore(core): remove import cycle or loop
