+++ mobile-app/app/screens/AppNavigator/screens/Balances/components/BottomSheetAssetSortList.tsx
@@ -5,7 +5,6 @@ import { Platform, TouchableOpacity } from 'react-native'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
 import { useThemeContext } from '@shared-contexts/ThemeProvider'
-import { BalancesSortType } from '../BalancesScreen'
 
 export interface BottomSheetAssetSortProps {
   headerLabel: string
@@ -13,6 +12,15 @@ export interface BottomSheetAssetSortProps {
   onButtonPress: (item: BalancesSortType) => void
   modifiedDenominationCurrency: string
   selectedAssetSortType: BalancesSortType
+}
+
+export enum BalancesSortType {
+  HighestDenominationValue = 'Highest denomination value',
+  LowestDenominationValue = 'Lowest denomination value',
+  HighestTokenAmount = 'Highest token amount',
+  LowestTokenAmount = 'Lowest token amount',
+  AtoZ = 'A to Z',
+  ZtoA = 'Z to A'
 }
 
 export const BottomSheetAssetSortList = ({
